feat(DialogData): Class that holds information about Dialogs regarding to tree needs
This class acts like a DialogNode for the tree. This is where the graph stores data like position of the node and the connections it has. As it replaces what previously was done with Dialog, there is one DialogData assigned to every node, they do not longer share them. Additionaly, it stores wether if the dialog should be played in runtime, as well as the possibleNextDialogs