feat(DialogGraphView): PopulateView working with main and unfinished branches
The function consists in generating the first node and a while cycle that generates nodes connected to it and stores in a list the ones that have other nodes connected to them. Then picks the first one in the list and repeats. When that is done, does the same with the DialogDatas in dialogTree.UnusedNodes to generate the whole working space as it was left