fix(DialogConnection): Change detection
Previously, the script did a lot of unnecesary loops. Now, it firstly checks for a different lenght between the two arrays, and if its the case, it obviously needs to be Updated. If not, the two arrays will have the exact same lenght, therefore we can check for differences in each position between the two arrays